+++ modules/admin-ui/src/main/resources/web/admin/apps/content-manager/js/app/wizard/page/LiveFormPanel.ts
@@ -56,9 +56,7 @@ module app.wizard.page {
     import RegionSelectEvent = api.liveedit.RegionSelectEvent;
     import ItemViewSelectedEvent = api.liveedit.ItemViewSelectedEvent;
     import ItemViewDeselectEvent = api.liveedit.ItemViewDeselectEvent;
-    import ComponentAddedEvent = api.liveedit.ComponentAddedEvent;
     import ComponentRemoveEvent = api.liveedit.ComponentRemoveEvent;
-    import ComponentResetEvent = api.liveedit.ComponentResetEvent;
     import ComponentDuplicateEvent = api.liveedit.ComponentDuplicateEvent;
 
     import Panel = api.ui.panel.Panel;
@@ -259,7 +257,6 @@ module app.wizard.page {
                     }
                 }
                 else if (api.ObjectHelper.iFrameSafeInstanceOf(event.getComponent(), ImageComponent)) {
-                    debugger;
                     if (event.getPropertyName() == ImageComponent.PROPERTY_IMAGE && !event.getComponent().isEmpty()) {
                         var componentView = this.pageView.getComponentViewByPath(event.getPath());
                         if (componentView) {
@@ -398,13 +395,6 @@ module app.wizard.page {
                 this.contextWindow.clearSelection();
             });
 
-            this.liveEditPage.onComponentReset((event: ComponentResetEvent) => {
-
-                if (!this.pageModel.isPageTemplate() && this.pageModel.getMode() == PageMode.AUTOMATIC) {
-                    this.pageModel.initializePageFromDefault(this);
-                }
-            });
-
             this.liveEditPage.onDraggingComponentViewStartedEvent((event: DraggingComponentViewStartedEvent) => {
 
                 if (this.contextWindow.isFloating() && this.contextWindow.isShown()) {
@@ -428,13 +418,6 @@ module app.wizard.page {
                 var toggler = this.contentWizardPanel.getContextWindowToggler();
                 if (this.contextWindow.isFloating() && !this.contextWindow.isShown() && toggler.isActive()) {
                     this.contextWindow.slideIn();
-                }
-            });
-
-            this.liveEditPage.onComponentAdded((event: ComponentAddedEvent) => {
-
-                if (!this.pageModel.isPageTemplate() && this.pageModel.getMode() == PageMode.AUTOMATIC) {
-                    this.pageModel.initializePageFromDefault(this);
                 }
             });