+++ modules/admin-ui/src/main/resources/web/admin/common/js/data/PropertyArray.ts
@@ -2,7 +2,7 @@ module api.data {
 
     export class PropertyArray implements api.Equitable {
 
-        public debug: boolean = false;
+        public static debug: boolean = false;
 
         private tree: PropertyTree;
 
@@ -22,12 +22,33 @@ module api.data {
 
         private propertyValueChangedListeners: {(event: PropertyValueChangedEvent):void}[] = [];
 
+        private propertyAddedEventHandler: (event: PropertyAddedEvent) => void;
+
+        private propertyRemovedEventHandler: (event: PropertyRemovedEvent) => void;
+
+        private propertyIndexChangedEventHandler: (event: PropertyIndexChangedEvent) => void;
+
+        private propertyValueChangedEventHandler: (event: PropertyValueChangedEvent) => void;
+
         constructor(builder: PropertyArrayBuilder) {
             this.tree = builder.parent.getTree();
             this.parent = builder.parent;
             this.name = builder.name;
             this.type = builder.type;
             this.array = [];
+
+            this.propertyAddedEventHandler = (event) => {
+                this.forwardPropertyAddedEvent(event)
+            };
+            this.propertyRemovedEventHandler = (event) => {
+                this.forwardPropertyRemovedEvent(event)
+            };
+            this.propertyIndexChangedEventHandler = (event) => {
+                this.forwardPropertyIndexChangedEvent(event)
+            };
+            this.propertyValueChangedEventHandler = (event) => {
+                this.forwardPropertyValueChangedEvent(event)
+            };
         }
 
         forEach(callBack: {(property: Property, index: number): void;}) {
@@ -190,6 +211,7 @@ module api.data {
 
             this.notifyPropertyRemoved(propertyToRemove);
             this.unregisterPropertyListeners(propertyToRemove);
+            propertyToRemove.detach();
         }
 
         public exists(index: number): boolean {
@@ -264,7 +286,7 @@ module api.data {
         }
 
         private registerPropertyListeners(property: Property) {
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].registerPropertyListeners: " +
                               property.getPath().toString());
@@ -275,24 +297,13 @@ module api.data {
                 this.registerPropertySetListeners(property.getSet());
             }
 
-            if (this.forwardPropertyIndexChangedEvent.bind) {
-                property.onPropertyIndexChanged(this.forwardPropertyIndexChangedEvent.bind(this));
-                property.onPropertyValueChanged(this.forwardPropertyValueChangedEvent.bind(this));
-            }
-            else {
-                // PhantomJS does not support bind
-                property.onPropertyIndexChanged((event) => {
-                    this.forwardPropertyIndexChangedEvent(event);
-                });
-                property.onPropertyValueChanged((event) => {
-                    this.forwardPropertyValueChangedEvent(event);
-                });
-            }
+            property.onPropertyIndexChanged(this.propertyIndexChangedEventHandler);
+            property.onPropertyValueChanged(this.propertyValueChangedEventHandler);
         }
 
         private unregisterPropertyListeners(property: Property) {
-            property.unPropertyIndexChanged(this.forwardPropertyIndexChangedEvent);
-            property.unPropertyValueChanged(this.forwardPropertyValueChangedEvent);
+            property.unPropertyIndexChanged(this.propertyIndexChangedEventHandler);
+            property.unPropertyValueChanged(this.propertyValueChangedEventHandler);
 
             if (property.hasNonNullValue() && property.getType().equals(ValueTypes.DATA)) {
                 var propertySet = property.getSet();
@@ -305,43 +316,26 @@ module api.data {
          */
         registerPropertySetListeners(propertySet: PropertySet) {
 
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].registerPropertySetListeners: " +
                               propertySet.getPropertyPath().toString());
             }
 
-            if (this.forwardPropertyAddedEvent.bind) {
-                propertySet.onPropertyAdded(this.forwardPropertyAddedEvent.bind(this));
-                propertySet.onPropertyRemoved(this.forwardPropertyRemovedEvent.bind(this));
-                propertySet.onPropertyIndexChanged(this.forwardPropertyIndexChangedEvent.bind(this));
-                propertySet.onPropertyValueChanged(this.forwardPropertyValueChangedEvent.bind(this));
-            }
-            else {
-                // PhantomJS does not support bind
-                propertySet.onPropertyAdded((event) => {
-                    this.forwardPropertyAddedEvent(event);
-                });
-                propertySet.onPropertyRemoved((event) => {
-                    this.forwardPropertyRemovedEvent(event);
-                });
-                propertySet.onPropertyIndexChanged((event) => {
-                    this.forwardPropertyIndexChangedEvent(event);
-                });
-                propertySet.onPropertyValueChanged((event) => {
-                    this.forwardPropertyValueChangedEvent(event);
-                });
-            }
+            propertySet.onPropertyAdded(this.propertyAddedEventHandler);
+            propertySet.onPropertyRemoved(this.propertyRemovedEventHandler);
+            propertySet.onPropertyIndexChanged(this.propertyIndexChangedEventHandler);
+            propertySet.onPropertyValueChanged(this.propertyValueChangedEventHandler);
         }
 
         /**
          * Package protected. Not to be used outside module.
          */
         unregisterPropertySetListeners(propertySet: PropertySet) {
-            propertySet.unPropertyAdded(this.forwardPropertyAddedEvent);
-            propertySet.unPropertyRemoved(this.forwardPropertyRemovedEvent);
-            propertySet.unPropertyIndexChanged(this.forwardPropertyIndexChangedEvent);
-            propertySet.unPropertyValueChanged(this.forwardPropertyValueChangedEvent);
+            propertySet.unPropertyAdded(this.propertyAddedEventHandler);
+            propertySet.unPropertyRemoved(this.propertyRemovedEventHandler);
+            propertySet.unPropertyIndexChanged(this.propertyIndexChangedEventHandler);
+            propertySet.unPropertyValueChanged(this.propertyValueChangedEventHandler);
         }
 
         onPropertyAdded(listener: {(event: PropertyAddedEvent): void;}) {
@@ -355,7 +349,7 @@ module api.data {
 
         private notifyPropertyAdded(property: Property) {
             var event = new PropertyAddedEvent(property);
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].notifyPropertyAdded: " +
                               event.getPath().toString());
@@ -364,7 +358,7 @@ module api.data {
         }
 
         private forwardPropertyAddedEvent(event: PropertyAddedEvent) {
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].forwardPropertyAddedEvent: " +
                               event.getPath().toString());
@@ -383,7 +377,7 @@ module api.data {
 
         private notifyPropertyRemoved(property: Property) {
             var event = new PropertyRemovedEvent(property);
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].notifyPropertyRemoved: " +
                               event.getPath().toString());
@@ -392,7 +386,7 @@ module api.data {
         }
 
         private forwardPropertyRemovedEvent(event: PropertyRemovedEvent) {
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].forwardPropertyRemovedEvent: " +
                               event.getPath().toString());
@@ -410,7 +404,7 @@ module api.data {
         }
 
         private forwardPropertyIndexChangedEvent(event: PropertyIndexChangedEvent) {
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].forwardPropertyIndexChangedEvent: " +
                               event.getPath().toString());
@@ -428,7 +422,7 @@ module api.data {
         }
 
         private forwardPropertyValueChangedEvent(event: PropertyValueChangedEvent) {
-            if (this.debug) {
+            if (PropertyArray.debug) {
                 console.debug("PropertyArray[" + this.getParentPropertyPath().toString() + "." + this.getName() +
                               "].forwardPropertyValueChangedEvent: " +
                               event.getPath().toString());