+++ modules/portal-api/src/main/java/com/enonic/xp/portal/url/PortalUrlService.java
@@ -7,4 +7,6 @@ public interface PortalUrlService
     public String serviceUrl( ServiceUrlParams params );
 
     public String pageUrl( PageUrlParams params );
+
+    public String componentUrl( ComponentUrlParams params );
 }