+++ modules/portal-api/src/test/java/com/enonic/xp/portal/url/PageUrlParamsTest.java
@@ -5,9 +5,6 @@ import org.junit.Test;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
 
-import com.enonic.wem.api.content.ContentId;
-import com.enonic.wem.api.content.ContentPath;
-
 import static org.junit.Assert.*;
 
 public class PageUrlParamsTest
@@ -23,7 +20,7 @@ public class PageUrlParamsTest
         assertNull( params.getId() );
 
         params.id( "123456" );
-        assertEquals( ContentId.from( "123456" ), params.getId() );
+        assertEquals( "123456", params.getId() );
     }
 
     @Test
@@ -36,7 +33,7 @@ public class PageUrlParamsTest
         assertNull( params.getPath() );
 
         params.path( "/a/b" );
-        assertEquals( ContentPath.from( "/a/b" ), params.getPath() );
+        assertEquals( "/a/b", params.getPath() );
     }
 
     @Test
@@ -50,8 +47,8 @@ public class PageUrlParamsTest
         final PageUrlParams params = configure( new PageUrlParams() );
         params.setAsMap( map );
 
-        assertEquals( ContentId.from( "123456" ), params.getId() );
-        assertEquals( ContentPath.from( "/a/b" ), params.getPath() );
+        assertEquals( "123456", params.getId() );
+        assertEquals( "/a/b", params.getPath() );
         assertEquals( "{a=[1]}", params.getParams().toString() );
         assertEquals( "PageUrlParams{params={a=[1]}, id=123456, path=/a/b}", params.toString() );
     }