+++ modules/portal-impl/src/main/java/com/enonic/xp/portal/impl/url/ContentIdResolver.java
@@ -1,0 +1,72 @@
+package com.enonic.xp.portal.impl.url;
+
+import com.enonic.wem.api.content.ContentId;
+import com.enonic.wem.api.content.ContentPath;
+import com.enonic.wem.api.content.ContentService;
+import com.enonic.xp.portal.PortalContext;
+
+final class ContentIdResolver
+{
+    private PortalContext context;
+
+    private ContentId id;
+
+    private ContentPath path;
+
+    private ContentService contentService;
+
+    public ContentIdResolver context( final PortalContext context )
+    {
+        this.context = context;
+        return this;
+    }
+
+    public ContentIdResolver id( final String value )
+    {
+        this.id = value != null ? ContentId.from( value ) : null;
+        return this;
+    }
+
+    public ContentIdResolver path( final String value )
+    {
+        this.path = value != null ? ContentPath.from( value ) : null;
+        return this;
+    }
+
+    public ContentIdResolver contentService( final ContentService value )
+    {
+        this.contentService = value;
+        return this;
+    }
+
+    public ContentId resolve()
+    {
+        if ( this.id != null )
+        {
+            return this.id;
+        }
+
+        if ( this.path != null )
+        {
+            return resolveId( this.path );
+        }
+
+        return null;
+    }
+
+    private ContentPath resolvePath( final ContentPath path )
+    {
+        if ( path.isAbsolute() )
+        {
+            return path;
+        }
+
+        return ContentPath.from( this.context.getContentPath(), path );
+    }
+
+    private ContentId resolveId( final ContentPath path )
+    {
+        final ContentPath resolved = resolvePath( path );
+        return this.contentService.getByPath( resolved ).getId();
+    }
+}