+++ modules/portal-impl/src/main/java/com/enonic/xp/portal/impl/url/PortalUrlServiceImpl.java
@@ -6,6 +6,7 @@ import org.osgi.service.component.annotations.Reference;
 import com.enonic.wem.api.content.ContentService;
 import com.enonic.xp.portal.url.AbstractUrlParams;
 import com.enonic.xp.portal.url.AssetUrlParams;
+import com.enonic.xp.portal.url.ComponentUrlParams;
 import com.enonic.xp.portal.url.PageUrlParams;
 import com.enonic.xp.portal.url.PortalUrlService;
 import com.enonic.xp.portal.url.ServiceUrlParams;
@@ -32,6 +33,12 @@ public final class PortalUrlServiceImpl
     public String pageUrl( final PageUrlParams params )
     {
         return build( new PageUrlBuilder(), params );
+    }
+
+    @Override
+    public String componentUrl( final ComponentUrlParams params )
+    {
+        return build( new ComponentUrlBuilder(), params );
     }
 
     private <B extends PortalUrlBuilder<P>, P extends AbstractUrlParams> String build( final B builder, final P params )