+++ modules/portal-jslib/src/test/resources/com/enonic/xp/portal/jslib/impl/ComponentUrlHandlerTest.js
@@ -7,5 +7,6 @@ exports.createUrl = function () {
         }
     });
 
-    assert.assertEquals('/portal/stage/some/path/_/component/mycomp?a=1&b=1&b=2', result);
+    // NOTE: This is not the actual url. Only a mock representation.
+    assert.assertEquals('ComponentUrlParams{params={a=[1], b=[1, 2]}, component=mycomp}', result);
 };